moved imaging prefix to naming/tests/organise structure
some options classes moved to avoid circular imports
